renomeia contracts para documents e move de indexing para raiz
