+++ etl/tests/test_silver_contract.py
@@ -1,7 +1,7 @@
 from django.test import SimpleTestCase
 
 from etl.indexing.schema import SILVER_MAPPING, SILVER_PROPERTIES
-from etl.indexing.contracts import BronzeDocument, SilverDocument
+from etl.documents import BronzeDocument, SilverDocument
 from etl.transform.standardizer import DefaultStandardizer
