+++ etl/transform/merger.py
@@ -1,6 +1,6 @@
 import logging
 
-from etl.indexing.contracts import SilverDocument
+from etl.documents import SilverDocument
 from etl.transform.normalizers import as_list, normalize_name, scalar_or_list, unique
 from etl.pipeline.defaults import DocumentRules
